Large refactor of the engine to separate out responsibilities. application is now only responsible for initializing the game objects and the core device, window and renderer.
We have a renderer class that now handles and manages the swap chain and command buffers and begining and ending the writing to the command buffers per frame.

The simple render system class now handles creating the pipeline and the pipeline layout for a given render pass and handles the actual drawing commands and binding the pipeline for drawing.